# Add KVCache trans for long sequence && tuned comm for faster Addreduce