Transpose KVCache with env var "ENABLE_KV_TRANS" for long sequence
Tune addreduce between shm and ccl for faster comm

| workload | baseline-2s | opt-2s |
| --------------- | --------------- | --------------- |
| llama-2-13b bs24 in512 bf16 (SPR+HBM) | 350ms | 90ms |
| llama-2-70b bs32 in4096 bf16 (SPR) | 1300ms | 640ms |